refactor: split metrics into submodules
