+++ src/tnfr/metrics/__init__.py
@@ -1,0 +1,46 @@
+from .core import (
+    register_metrics_callbacks,
+    Tg_global,
+    Tg_by_node,
+    latency_series,
+    glifogram_series,
+    glyph_top,
+    glyph_dwell_stats,
+    _tg_state,
+    _update_tg,
+    _update_latency_index,
+    _update_epi_support,
+    _metrics_step,
+)
+
+from .coherence import (
+    coherence_matrix,
+    local_phase_sync_weighted,
+    register_coherence_callbacks,
+)
+from .diagnosis import (
+    register_diagnosis_callbacks,
+    dissonance_events,
+)
+from .export import export_history
+
+__all__ = [
+    "register_metrics_callbacks",
+    "Tg_global",
+    "Tg_by_node",
+    "latency_series",
+    "glifogram_series",
+    "glyph_top",
+    "glyph_dwell_stats",
+    "_tg_state",
+    "_update_tg",
+    "_update_latency_index",
+    "_update_epi_support",
+    "_metrics_step",
+    "coherence_matrix",
+    "local_phase_sync_weighted",
+    "register_coherence_callbacks",
+    "register_diagnosis_callbacks",
+    "dissonance_events",
+    "export_history",
+]
